+++ src/Definitions/Shipper.php
@@ -227,6 +227,20 @@ final class Shipper
     public const JAPO = 'japo';
 
     /**
+     * Liftago
+     *
+     * @var string
+     */
+    public const LIFTAGO = 'liftago';
+
+    /**
+     * Magyar posta
+     *
+     * @var string
+     */
+    public const MAGYARPOSTA = 'magyarposta';
+
+    /**
      * All supported shipper services.
      *
      * @return array<string>
@@ -265,6 +279,8 @@ final class Shipper
             self::DBSCHENKER,
             self::AIRWAY,
             self::JAPO,
+            self::LIFTAGO,
+            self::MAGYARPOSTA,
         ];
     }