+++ tests/Integration/Balikobot/GetManipulationUnitsMethodTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Inspirum\Balikobot\Tests\Integration\Balikobot;
 
+use Inspirum\Balikobot\Contracts\ExceptionInterface;
 use Inspirum\Balikobot\Definitions\Shipper;
-use Inspirum\Balikobot\Exceptions\BadRequestException;
 use function count;
 use function is_array;
 use function is_int;
@@ -42,10 +42,13 @@ class GetManipulationUnitsMethodTest extends AbstractBalikobotTestCase
 
     public function testInvalidRequest(): void
     {
-        $this->expectException(BadRequestException::class);
-
         $service = $this->newBalikobot();
 
-        $service->getManipulationUnits(Shipper::CP);
+        try {
+            $service->getManipulationUnits(Shipper::CP);
+            self::fail('MANIPULATIONUNITS request should thrown exception');
+        } catch (ExceptionInterface $exception) {
+            self::assertEquals(501, $exception->getStatusCode());
+        }
     }
 }